se logra conectar back con front y se imprime en pantalla los posts

+++ frontend/src/api/axios.js
@@ -1,42 +1,51 @@
 import axios from "axios";
 
-const api = axios.create({
+const instance = axios.create({
   baseURL: "http://127.0.0.1:8000/api/",
 });
 
-api.interceptors.request.use((config) => {
+instance.interceptors.request.use((config) => {
   const token = localStorage.getItem("access");
+
   if (token) {
     config.headers.Authorization = `Bearer ${token}`;
   }
+
   return config;
 });
 
-api.interceptors.response.use(
+instance.interceptors.response.use(
   (response) => response,
   async (error) => {
     const originalRequest = error.config;
 
-    if (error.response?.status === 401 && !originalRequest._retry) {
+    if (
+      error.response?.status === 401 &&
+      !originalRequest._retry &&
+      !originalRequest.url.includes("/login/")
+    ) {
       originalRequest._retry = true;
 
       const refresh = localStorage.getItem("refresh");
 
       if (!refresh) {
+        localStorage.clear();
         window.location.href = "/login";
-        return;
+        return Promise.reject(error);
       }
 
       try {
         const res = await axios.post(
-          "http://127.0.0.1:8000/api/token/refresh/",
+          "http://127.0.0.1:8000/api/refresh/",
           { refresh }
         );
 
         localStorage.setItem("access", res.data.access);
 
-        originalRequest.headers.Authorization = `Bearer ${res.data.access}`;
-        return api(originalRequest);
+        originalRequest.headers.Authorization =
+          "Bearer " + res.data.access;
+
+        return instance(originalRequest);
       } catch (err) {
         localStorage.clear();
         window.location.href = "/login";
@@ -46,5 +55,4 @@ api.interceptors.response.use(
     return Promise.reject(error);
   }
 );
-
-export default api;
+export default instance;
